Wow that was dumb.. Moved logging under Services namespace and split classes into their own files. Some PropertyStore tests as well too
