+++ source/Xeno.ToolsHub.SampleAssmAddin/Handlers/OnStartHandler.cs
@@ -6,11 +6,11 @@
  *  SampleAssembly OnStart extension
  */
 
-using Xeno.ToolsHub.Config;
-using Xeno.ToolsHub.ExtensionModel;
-
 namespace Xeno.ToolsHub.SampleAssmAddin.Handlers
 {
+  using Xeno.ToolsHub.ExtensionModel;
+  using Xeno.ToolsHub.Services.Logging;
+
   [Mono.Addins.Extension(NodeName = ExtensionName.OnStartupAddin, Path = ExtensionPath.OnStartup)]
   public class OnStartHandler : ExtensionModel.IOnStartupExtension
   {
