+++ source/Xeno.ToolsHub.Tests/Xeno.ToolsHub.Tests.csproj
@@ -58,12 +58,12 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SystemTests\HelperTests.cs" />
-    <Compile Include="SystemTests\PropertyTests\PropertiesTests.cs" />
+    <Compile Include="SystemTests\PropertiesTests.cs" />
     <Compile Include="Prototypes\PropertySingleton\IPropertyService.cs" />
     <Compile Include="Prototypes\PropertySingleton\Properties.cs" />
     <Compile Include="Prototypes\PropertySingleton\PropertyService.cs" />
     <Compile Include="Helpers\PropertyHelpers.cs" />
-    <Compile Include="SystemTests\PropertyTests\PropertiesStoreTests.cs" />
+    <Compile Include="SystemTests\PropertiesStoreTests.cs" />
     <Compile Include="Helpers\TestHelpers.cs" />
   </ItemGroup>
   <ItemGroup>
